requirements.txt
fpdf==1.7.2
pandas==2.2.3
Pillow==11.1.0
streamlit==1.41.1
